s03e03/requirements.txt
python-dotenv
requests
langchain
langchain-core
langchain-openai
tiktoken
